Added a method in Solution to get a certain planning with a given route
Finished first iteration of GeneticRouteStrategy.
In this Strategy a random route will be selected in a random cluster.
The specified cluster must have enough routes ( > 0) and the route in the target cluster must have enough orders ( > 2, 1 is the 0 order so do not count that one and it need at least to elements to swap), otherwise a different route/cluster will be selected.
Then it will randomly calculate the number of slices it is going to make and where to make them.
When this is done it will randomly reassign the slices to a new route and this route will be added to the solution