Initial refactor of changing OrdersCounter to OrdersTracker.
Route will automatically add and remove orderoccurences when orders are added or removed.

Solution will track if routes are planned in or removed from planning - HOWEVER it does not function properly yet.

Solution still relies on OrdersTracker for calculating the SolutionScore. This is NOT correct, considering routes don't neccesarily have to be planned in.

Order is no longer responsible for knowing of what routes it is a part of.

Did some renaming across the board.